refactor the RMID allocator to rmid_allocator.{h,c}
