+++ .vscode/c_cpp_properties.json
@@ -12,7 +12,7 @@
                 "/usr/src/linux-headers-6.8.0-52-generic/arch/x86/include/generated/uapi",
                 "/usr/src/linux-headers-6.8.0-52-generic"
             ],
-            "defines": ["__KERNEL__", "MODULE"],
+            "defines": ["__KERNEL__", "MODULE", "CONFIG_X86_CPU_RESCTRL"],
             "compilerPath": "/usr/bin/gcc",
             "cStandard": "c11",
             "intelliSenseMode": "linux-gcc-x64"
@@ -29,7 +29,7 @@
                 "/usr/src/linux-headers-$(uname -r)/arch/x86/include/generated/uapi",
                 "/usr/src/linux-headers-$(uname -r)"
             ],
-            "defines": ["__KERNEL__", "MODULE"],
+            "defines": ["__KERNEL__", "MODULE", "CONFIG_X86_CPU_RESCTRL"],
             "compilerPath": "/usr/bin/gcc",
             "cStandard": "c11",
             "intelliSenseMode": "linux-gcc-x64"
