+++ module/collector.h
@@ -1,0 +1,6 @@
+#ifndef COLLECTOR_H
+#define COLLECTOR_H
+
+#define LOG_PREFIX "Memory Collector: "
+
+#endif
