+++ module/memory_collector.c
@@ -12,49 +12,26 @@
 #include "rdt.h"
 #include <linux/workqueue.h>
 #include "procfs.h"
+#include "rmid_allocator.h"
+#include "collector.h"
 
 MODULE_LICENSE("GPL");
 MODULE_AUTHOR("Memory Collector Project");
 MODULE_DESCRIPTION("Memory subsystem monitoring for Kubernetes");
 MODULE_VERSION("1.0");
 
-#define LOG_PREFIX "Memory Collector: "
-
 // Define the tracepoint
 #define CREATE_TRACE_POINTS
 #include "memory_collector_trace.h"
 
 #define EMULATED_MAX_RMID 512
-#define RMID_INVALID 0
 #define CLOSID_CATCHALL 0
 
-// RMID allocation structure
-struct rmid_info {
-    struct list_head list;  // For free list
-    u32 rmid;
-    char comm[TASK_COMM_LEN];  // Name of task leader
-    pid_t tgid;  // Thread group ID (process ID)
-};
-
-struct rmid_alloc {
-    spinlock_t lock;  // Protects all fields
-    struct list_head free_list;  // List of free RMIDs
-    u32 max_rmid;  // Minimum of max_rmid across all CPUs
-    struct rmid_info *rmids;  // Array of RMID info, indexed by RMID
-    bool hardware_support;  // true if RDT hardware support is detected
-};
-
-static struct rmid_alloc rmid_allocator;
-
 // Forward declarations of new functions
-static void rmid_free(u32 rmid);
-static u32 _rmid_alloc(const char *comm, pid_t tgid);
-static int init_rmid_allocator(void);
-static void cleanup_rmid_allocator(void);
-static void assign_rmid_to_task(struct task_struct *task);
 static void assign_rmids_to_leaders(void);
 static void propagate_leader_rmids(void);
 static void reset_cpu_rmid(void *info);
+static int detect_and_init_rmid_allocator(void);
 
 struct cpu_state {
     struct hrtimer timer;
@@ -70,6 +47,8 @@ static enum hrtimer_restart timer_fn(struct hrtimer *timer);
 // Add global cpu_works declaration after other global variables
 static struct work_struct __percpu *cpu_works;
 static struct workqueue_struct *collector_wq;
+
+static bool rdt_hardware_support = false;
 
 static void collect_sample_on_current_cpu(bool is_context_switch)
 {
@@ -92,7 +71,7 @@ static void probe_sched_switch(void *data,
     collect_sample_on_current_cpu(true);
 
     // Update RMID if it's changing and we have hardware support
-    if (prev->rmid != next->rmid && rmid_allocator.hardware_support) {
+    if (prev->rmid != next->rmid && rdt_hardware_support) {
         rdt_write_rmid_closid(next->rmid, CLOSID_CATCHALL);
     }
 }
@@ -177,78 +156,6 @@ static void reset_all_task_rmids(void)
         }
     }
     rcu_read_unlock();
-}
-
-/*
- * Internal helper to allocate an RMID.
- * Caller must hold rmid_allocator.lock.
- * Returns 0 if no RMID is available (RMID 0 is reserved/invalid).
- */
-static u32 _rmid_alloc(const char *comm, pid_t tgid)
-{
-    struct rmid_info *info;
-    u32 rmid;
-
-    // Check if we have any free RMIDs
-    if (list_empty(&rmid_allocator.free_list)) {
-        return 0;  // RMID 0 is reserved/invalid
-    }
-
-    // Get the RMID that was freed the longest time ago
-    info = list_first_entry(&rmid_allocator.free_list, struct rmid_info, list);
-    list_del_init(&info->list);
-
-    // Update RMID info
-    strncpy(info->comm, comm, TASK_COMM_LEN - 1);
-    info->comm[TASK_COMM_LEN - 1] = '\0';
-    info->tgid = tgid;
-    rmid = info->rmid;
-
-    // Emit tracepoint for RMID allocation while holding the lock
-    trace_memory_collector_rmid_alloc(rmid, comm, tgid, ktime_get_ns());
-
-    return rmid;
-}
-
-static void assign_rmid_to_task(struct task_struct *task)
-{
-    struct task_struct *group_leader;
-    u32 rmid;
-    unsigned long flags;
-
-    if (!task)
-        return;
-
-    group_leader = task->group_leader;
-    if (!group_leader)
-        return;
-
-    // If this is not the group leader, just copy the leader's RMID
-    if (task != group_leader) {
-        task->rmid = group_leader->rmid;
-        return;
-    }
-
-    // First check without lock
-    if (group_leader->rmid)
-        return;  // Leader already has an RMID
-
-    // We do not assign RMIDs to kernel threads
-    if ((group_leader->mm == NULL) || (group_leader->flags & PF_KTHREAD))
-        return;
-
-    // No RMID assigned to leader, need to allocate one
-    spin_lock_irqsave(&rmid_allocator.lock, flags);
-
-    // Recheck after acquiring lock
-    if (!group_leader->rmid) {
-        // Allocate new RMID for the process
-        rmid = _rmid_alloc(group_leader->comm, group_leader->tgid);
-        group_leader->rmid = rmid;
-        // Note: if allocation fails, leader->rmid remains 0
-    }
-
-    spin_unlock_irqrestore(&rmid_allocator.lock, flags);
 }
 
 // Tracepoint probes for process lifecycle events
@@ -339,35 +246,6 @@ static int init_tracepoints(void)
     return 0;
 }
 
-extern void dump_existing_rmids(void);
-
-// Make dump_existing_rmids available to procfs.c
-void dump_existing_rmids(void)
-{
-    unsigned long flags;
-    u32 i;
-    struct rmid_info *info;
-
-    for (i = 1; i <= rmid_allocator.max_rmid; i++) {
-        // Lock for each element to avoid starving the write path
-        spin_lock_irqsave(&rmid_allocator.lock, flags);
-        
-        info = &rmid_allocator.rmids[i];
-
-        // Only emit tracepoint if RMID is in use (not on free list)
-        if (list_empty(&info->list)) {
-            trace_memory_collector_rmid_existing(
-                info->rmid,
-                info->comm,
-                info->tgid,
-                ktime_get_ns()
-            );
-        }
-        
-        spin_unlock_irqrestore(&rmid_allocator.lock, flags);
-    }
-}
-
 static int __init memory_collector_init(void)
 {
     int ret;
@@ -419,7 +297,7 @@ static int __init memory_collector_init(void)
     flush_workqueue(collector_wq);
 
     // Initialize RMID allocator after all CPUs are initialized
-    ret = init_rmid_allocator();
+    ret = detect_and_init_rmid_allocator();
     if (ret) {
         pr_err(LOG_PREFIX "Failed to initialize RMID allocator\n");
         goto err_cleanup_cpu_states;
@@ -490,7 +368,7 @@ static void __exit memory_collector_exit(void)
     tracepoint_synchronize_unregister();
 
     // Reset RMID to 0 on all CPUs if we have hardware support
-    if (rmid_allocator.hardware_support) {
+    if (rdt_hardware_support) {
         on_each_cpu(reset_cpu_rmid, NULL, 1);
     }
 
@@ -507,7 +385,7 @@ static void __exit memory_collector_exit(void)
 }
 
 // RMID allocation and initialization functions
-static int init_rmid_allocator(void)
+static int detect_and_init_rmid_allocator(void)
 {
     int cpu;
     u32 min_max_rmid = U32_MAX;
@@ -522,62 +400,15 @@ static int init_rmid_allocator(void)
 
     if (min_max_rmid == U32_MAX || min_max_rmid == 0) {
         min_max_rmid = EMULATED_MAX_RMID;
-        rmid_allocator.hardware_support = false;
+        rdt_hardware_support = false;
         pr_info(LOG_PREFIX "Using emulated RMIDs (max=%d)\n", EMULATED_MAX_RMID);
     } else {
-        rmid_allocator.hardware_support = true;
+        rdt_hardware_support = true;
         pr_info(LOG_PREFIX "Using hardware RMIDs (max=%d)\n", min_max_rmid);
     }
 
     // Initialize allocator structure with spinlock
-    spin_lock_init(&rmid_allocator.lock);
-    INIT_LIST_HEAD(&rmid_allocator.free_list);
-    rmid_allocator.max_rmid = min_max_rmid;
-
-    // Allocate array of RMID info structures
-    rmid_allocator.rmids = kzalloc(sizeof(struct rmid_info) * (min_max_rmid + 1), GFP_KERNEL);
-    if (!rmid_allocator.rmids) {
-        pr_err(LOG_PREFIX "Failed to allocate RMID info array\n");
-        return -ENOMEM;
-    }
-
-    // Initialize all RMIDs (skip RMID 0 as it's reserved)
-    for (u32 i = 0; i <= min_max_rmid; i++) {
-        INIT_LIST_HEAD(&rmid_allocator.rmids[i].list);
-        rmid_allocator.rmids[i].rmid = i;
-        rmid_allocator.rmids[i].tgid = 0;
-        if (i != RMID_INVALID) {  // Don't add RMID 0 to free list
-            list_add_tail(&rmid_allocator.rmids[i].list, &rmid_allocator.free_list);
-        }
-    }
-
-    return 0;
-}
-
-static void cleanup_rmid_allocator(void)
-{
-    kfree(rmid_allocator.rmids);
-    rmid_allocator.rmids = NULL;
-}
-
-static void rmid_free(u32 rmid)
-{
-    unsigned long flags;
-    struct rmid_info *info;
-
-    if (rmid == RMID_INVALID || rmid > rmid_allocator.max_rmid)
-        return;
-
-    spin_lock_irqsave(&rmid_allocator.lock, flags);
-
-    info = &rmid_allocator.rmids[rmid];
-    info->tgid = 0;
-    list_add_tail(&info->list, &rmid_allocator.free_list);
-
-    // Emit tracepoint for RMID deallocation while holding the lock
-    trace_memory_collector_rmid_free(rmid, ktime_get_ns());
-
-    spin_unlock_irqrestore(&rmid_allocator.lock, flags);
+    return init_rmid_allocator(min_max_rmid);
 }
 
 /*
